Alteração(Controllers): Recebendo ServerResquestInterface e Retornando ResponseInterface

+++ src/Controller/RemoveVideoController.php
@@ -2,9 +2,12 @@
 
 namespace HackbartPR\Controller;
 
+use Nyholm\Psr7\Response;
 use HackbartPR\Utils\Message;
 use HackbartPR\Interfaces\Controller;
 use HackbartPR\Repository\PDOVideoRepository;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 
 class RemoveVideoController implements Controller
 {
@@ -18,25 +21,36 @@ class RemoveVideoController implements Controller
         $this->repository = $repository;
     }
 
-    public function processRequest(): void
+    public function processRequest(ServerRequestInterface $request): ResponseInterface
     {
-        [$id] = $this->validation();
+        [$id, $validation] = $this->validation($request);
+
+        if (!$validation) {
+            return new Response(404, ['Location' => '/']);
+        }
+
         $result = $this->repository->remove($id);
 
         if ($result) {
             $this->create(self::REMOVE_SUCCESS);
+            return new Response(204, ['Location' => '/']);
         } else {
             $this->create(self::REMOVE_FAIL);
+            return new Response(400, ['Location' => '/']);
         }
     }
 
-    private function validation(): array
-    {
-        if (!isset($_GET['id'])) {
+    private function validation(ServerRequestInterface $request): array
+    {   
+        $body = $request->getParsedBody();
+
+        if (empty($body['id'])) {
             $this->create(self::NOT_FOUND);
+            $validation = false;
         }
 
-        $id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);        
-        return [$id];        
+        $id = filter_var($body['id'], FILTER_VALIDATE_INT);        
+
+        return [$id, $validation];        
     }
 }
